refactor(blorktools): state hoist
