+++ packages/blorktools/src/asset_debugger/util/animation/playback/css3d-reversal-helper.js
@@ -1,4 +1,4 @@
-import { animationStack, reverseAnimationFrameId } from "../state/css3d-state";
+import { animationStack, reverseAnimationFrameId } from "../../state/css3d-state";
 
 /**
  * Pop and play the next animation from the stack in reverse
